security: "secure" directory creation never repaired an existing mode, and the Rust action server had the C++ duplicate-goal bug too
Seven findings from the open backlog. One was found by a test failing on my own
change, and it is the most interesting of the batch.

1. create_dir_secure silently accepted a world-readable directory.

   `DirBuilder::mode(0o700)` applies ONLY to directories the call creates. When
   the path already exists the call succeeds and leaves the existing mode
   untouched. So once any looser path created the tree first — a bare
   `create_dir_all` gets 0o775 under the usual 0o002 umask — every subsequent
   `create_dir_secure` accepted it, forever, from a function whose entire
   purpose is to prevent exactly that. For `shm_nodes_dir()` that means the
   robot's node inventory readable by any local user.

   Repairs by MASKING, never assigning, and only on directories this user owns:
   a path already stricter than 0o700 keeps its stricter mode. Assigning
   outright would ADD owner-write to a 0o444 directory — the same bug I
   introduced once already in this area and that an existing test caught.

   Found because `test_presence_dir_created_with_mode_0700` started failing
   after change 2 below. It was passing only because the directory happened to
   be 0o700 on this machine; the defect it documents is pre-existing and
   independent of anything on this branch.

2. Two writers bypassed the hardened creator entirely (F3, SHM-4).

   `write_topic_meta` and horus_net's presence receiver both used bare
   `create_dir_all`, skipping the ownership gate AND the mode. Whichever code
   path happened to touch the tree first decided its permissions — and the
   horus_net one runs on the RECEIVE path, so a remote peer's presence
   announcement was enough to create the tree loosely before any local process
   hardened it. Both now go through `create_shm_dir_all`.

3. The ownership gate was write-path-only (F3).

   A read-only tool — `horus topic list`, the monitor, anything that never
   publishes — walked an attacker pre-created `/dev/shm/horus_<ns>` without
   noticing it belonged to someone else. Individual `.meta` names were already
   validated, but the operator was still shown another user's topic inventory
   as their own. `find_topics` now warns once. Warns rather than refuses
   deliberately: this is the read path, and a diagnostic tool that silently
   returns nothing is worse than one that says why. Write paths still fail
   closed.

4. The Rust action server had the same duplicate-goal defect as the C++ one
   (ACT-3).

   Fixed in horus_cpp two commits ago; horus_core has the identical shape and
   was missed — the same "guarded one call site, not its sibling" pattern that
   has now appeared five times on this branch. A duplicate goal_id replaced the
   live goal's GoalState, including the cancel/preempt Arcs its running callback
   polls, and dropped its JoinHandle, detaching the thread from the map `drain()`
   joins at shutdown.

   The rationale differs from the C++ case and the comment says so rather than
   being copied across: `GoalId` here wraps a v4 UUID, so accidental collision
   is not realistic — the vector is a publisher that reuses or replays an id on
   a topic any local process can write.

5. Replay allocated from an untrusted count (RR-1). `total_ticks` comes out of
   a recording file and sized a Vec<u64> one entry per tick; a few bytes of
   JSON declaring 2^60 ticks asks for exabytes. Bounded at 100M (800 MB, over a
   day at 1 kHz).

6. Replay joined an untrusted path (RR-3). `node_recordings` values were joined
   verbatim, so `../../..` escaped the session directory and `add_replay` opened
   whatever it landed on. Every other path field from this file already went
   through `sanitize_path_component`; this one was missed.

7. Terminal escape injection into operator output (F4). Everything `horus topic
   echo` prints is publisher-controlled. The `PodText` arm already refused
   anything outside ASCII graphic + whitespace, but its siblings — PodFields,
   Json, and the header's type_name — did not, so the guard covered only the
   format least likely to carry structured text. An ESC in that data is not
   text, it is a command: `\x1b]0;...` rewrites the window title and `\x1b[2J`
   clears the scrollback an operator is reading during an incident. Now
   neutralised on every arm, C1 controls included, keeping tab and newline
   because the output is deliberately multi-line.

Verified during this pass and NOT changed, because they were already fixed and
merely unmarked in the findings file: F1 (cap_mask is validated, and data_len is
bounded), F2/SHM-3 (find_topics validates .meta names — 34e08943), MON-1 (the
project-local monitor binary is opt-in behind HORUS_ALLOW_LOCAL_PLUGINS),
ACT-1 (the preempt table is bounded).

The action-server and terminal tests were both confirmed to FAIL against the
pre-fix code ("the live goal's cancel flag was REPLACED"), and the directory
repair was verified by setting a live tree to 0o775 and watching it come back
0o700.

horus_core 2245/2245, horus_sys 176/176, horus_net 317/317,
horus_manager 3099/3099, horus_cpp 136/136, horus_types 171/171.

+++ horus_py/src/pointcloud.rs
@@ -117,7 +117,9 @@ impl PyPointCloud {
         ];
         // See PyImage::__array_interface__: point_count/fields_per_point come off
         // the wire and the export hands numpy a raw pointer into the pool slot.
-        match (this.inner.point_count() as u64)
+        match this
+            .inner
+            .point_count()
             .checked_mul(this.inner.fields_per_point() as u64)
         {
             Some(n) if tensor.element_count_fits(n) => {}
